Fix bug recording when no file in Windows
Windows was allowing "gtm" to be called without a filename when the
editor was opened without a project folder and when a new file was
in the main editor screen. The "filename" parameter which was being
checked for having something other than "" (!= "") failed because
filename was passed in undefined in that particular situation.
Only happened in Windows, but a good thing to check anyway.